enhance/merge-config
