--- .env.example
@@ -1,7 +1,0 @@
-WAPPIN_BASE_URL=
-WAPPIN_CLIENT_ID=
-WAPPIN_TIMEOUT=15
-
-WAPPIN_CACHE_DRIVER=redis
-WAPPIN_CACHE_HOST=127.0.0.1
-WAPPIN_CACHE_PORT=6379
