+++ .gitignore
@@ -1,1 +1,3 @@
 .env
+/vendor
+wappin
