+++ go.mod
@@ -1,13 +1,13 @@
 module github.com/flip-id/wappin
 
-go 1.13
+go 1.14
 
 require (
-	github.com/eko/gocache v1.0.0
-	github.com/go-redis/redis/v7 v7.0.0-beta.5
-	github.com/go-resty/resty/v2 v2.1.0
-	github.com/jarcoal/httpmock v1.0.4
-	github.com/joho/godotenv v1.3.0
-	github.com/sirupsen/logrus v1.4.2
-	github.com/stretchr/testify v1.4.0
+	github.com/go-resty/resty/v2 v2.6.0
+	github.com/jarcoal/httpmock v1.0.8
+	github.com/joho/godotenv v1.4.0
+	github.com/kr/pretty v0.1.0 // indirect
+	github.com/sirupsen/logrus v1.8.1
+	github.com/stretchr/testify v1.7.0
+	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
 )
