+++ wappin.go
@@ -3,13 +3,21 @@ package wappin
 import (
 	"encoding/json"
 	"errors"
+	"time"
 
+	"github.com/go-resty/resty/v2"
 	log "github.com/sirupsen/logrus"
 )
 
-const SendHsmEndpoint = "/v1/message/do-send-hsm"
+const SEND_HSM_ENDPOINT = "/v1/message/do-send-hsm"
+const TOKEN_ENDPOINT = "/v1/token/get"
+const CONNECTION_TIME_OUT = 15
+
+var client = resty.New().SetTimeout(time.Second * time.Duration(CONNECTION_TIME_OUT))
 
 type Config struct {
+	BaseUrl   string
+	ClientId  string
 	ProjectId string
 	SecretKey string
 	ClientKey string
@@ -48,26 +56,23 @@ type CallbackData struct {
 	Timestamp      string `json:"timestamp"`
 }
 
+type AccessToken struct {
+	ClientId string
+	Status   string `json:"status"`
+	Message  string `json:"message"`
+	Data     struct {
+		AccessToken     string `json:"access_token"`
+		ExpiredDatetime string `json:"expired_datetime"`
+		TokenType       string `json:"token_type"`
+	} `json:"data"`
+}
+
 // Create sender object
 func New(config Config) *Sender {
 	return &Sender{Config: config}
 }
 
-// Set authorization token
-func (s *Sender) setToken() error {
-	accessToken, err := getAccessToken(s.Config.SecretKey)
-
-	if err != nil {
-		return err
-	}
-
-	s.AccessToken = accessToken
-
-	return nil
-}
-
 func (s *Sender) SendMessage(reqMsg interface{}) (res ResMessage, err error) {
-	err = s.setToken()
 
 	if err != nil {
 		return ResMessage{
@@ -89,14 +94,14 @@ func (s *Sender) SendMessage(reqMsg interface{}) (res ResMessage, err error) {
 
 // Send Whatsapp message
 func (s *Sender) sendWaMessage(req ReqWaMessage) (ResMessage, error) {
-	res, err := s.postToWappin(SendHsmEndpoint, req)
+	res, err := s.postToWappin(SEND_HSM_ENDPOINT, req)
 
 	return res, err
 }
 
 // Post request to Wappin service
 func (s *Sender) postToWappin(endpoint string, body interface{}) (ResMessage, error) {
-	url := baseUrl + endpoint
+	url := s.Config.BaseUrl + endpoint
 	res, err := client.R().SetBody(body).SetAuthToken(s.AccessToken.Data.AccessToken).Post(url)
 	resMessage := ResMessage{}
 
@@ -121,5 +126,32 @@ func (s *Sender) postToWappin(endpoint string, body interface{}) (ResMessage, er
 		return resMessage, err
 	}
 
-	return resMessage, err
+	return resMessage, nil
+}
+
+func (s *Sender) GenerateAccessToken() (AccessToken, error) {
+	url := s.Config.BaseUrl + TOKEN_ENDPOINT
+	accessToken := AccessToken{}
+	res, err := client.R().SetBasicAuth(clientId, s.Config.SecretKey).Post(url)
+
+	if err != nil {
+		return accessToken, err
+	}
+
+	if err := json.Unmarshal(res.Body(), &accessToken); err != nil {
+		return accessToken, err
+	}
+
+	if accessToken.Status != "200" {
+		log.WithFields(log.Fields{
+			"msg": "Failed to get token",
+			"res": res,
+		}).Error()
+		return accessToken, errors.New(accessToken.Message)
+	}
+
+	// Set cache
+	//err = setAccessToken(s.Config.SecretKey, &accessToken)
+
+	return accessToken, err
 }
