+++ wappin_test.go
@@ -13,7 +13,7 @@ func TestSendNotificationHSM(t *testing.T) {
 
 	fixture := `{ "message_id": "id-123", "status": "200", "message": "Success" }`
 	responder := httpmock.NewStringResponder(200, fixture)
-	fakeUrl := baseUrl + SendHsmEndpoint
+	fakeUrl := baseUrl + SEND_HSM_ENDPOINT
 	httpmock.RegisterResponder("POST", fakeUrl, responder)
 
 	config := Config{
@@ -47,7 +47,7 @@ func TestFailSendNotificationHSM(t *testing.T) {
 
 	fixture := `{ "message_id": "id-124", "status": "600", "message": "Not delivered, Contact validate Failed" }`
 	responder := httpmock.NewStringResponder(200, fixture)
-	fakeUrl := baseUrl + SendHsmEndpoint
+	fakeUrl := baseUrl + SEND_HSM_ENDPOINT
 	httpmock.RegisterResponder("POST", fakeUrl, responder)
 
 	config := Config{
@@ -95,6 +95,6 @@ func TestInvalidRequestFormat(t *testing.T) {
 func mockGetAccessToken() {
 	fixture := `{ "status": "200", "message": "Success", "data": { "access_token": "677b800f9b694f98bb9db6edb18336743a3f416cadff1953a59190f309220936", "expired_datetime": "2020-12-28 10:20:23", "token_type": "Bearer" } }`
 	responder := httpmock.NewStringResponder(200, fixture)
-	fakeUrl := baseUrl + TokenEndpoint
+	fakeUrl := baseUrl + TOKEN_ENDPOINT
 	httpmock.RegisterResponder("POST", fakeUrl, responder)
 }
